Rename to git_log.py on feature "step" branch
These "step" branches are used to quickly swap between code when demonstrating.

This particular commit is not needed in master. It is here to update the "step"
branch so that when we switch between step branches, things are more clear.
Upstream, this filename is renamed to "git_log.py," so we will do it here as
well.